switch timestamp to u64

+++ src/models/builders.rs
@@ -45,7 +45,7 @@ impl Game {
             rated: seek.rated,
             game: game_type,
             clock,
-            created_at: Utc::now().timestamp_millis(),
+            created_at: Utc::now().timestamp_millis() as u64,
             first,
             second,
             start_pos: seek.start_pos,
